backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
openai==1.3.7
langchain==0.1.0
langgraph==0.0.20
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3 